نحوه استفاده از Custom Exception های ایجاد شده در CQRS

+++ HR_Management.Application/Features/LeaveRequests/Handlers/Commands/DeleteLeaveRequestCommandHandler.cs
@@ -3,6 +3,8 @@ using System.Threading.Tasks;
 using System.Threading;
 using HR_Management.Application.Features.LeaveRequests.Requests.Commands;
 using HR_Management.Application.Persistence.Contracts;
+using HR_Management.Application.Exceptions;
+using HR_Management.Domain;
 
 namespace HR_Management.Application.Features.LeaveRequests.Handlers.Commands
 {
@@ -19,6 +21,11 @@ namespace HR_Management.Application.Features.LeaveRequests.Handlers.Commands
         public async Task<Unit> Handle(DeleteLeaveRequestCommand request, CancellationToken cancellationToken)
         {
             var leaveRequest = await _leaveRequestRepository.Get(request.Id);
+
+            if (leaveRequest == null)
+            {
+                throw new NotFoundException(nameof(LeaveRequest), request.Id);
+            }
 
             await _leaveRequestRepository.Delete(leaveRequest);
             return Unit.Value;
